requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiofiles>=23.2.1
aioredis==2.0.1
aiosignal==1.4.0
alembic==1.13.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
bcrypt==4.2.0
bleach==6.1.0
cachetools==5.5.2
catboost>=1.2.5
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==43.0.1
cycler==0.12.1
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fairlearn==0.10.0
fastapi==0.104.1
fonttools==4.60.1
frozenlist==1.8.0
geoip2==4.8.0
google-api-core==2.25.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
gotrue==2.12.3
graphviz==0.21
greenlet==3.2.4
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
joblib==1.3.2
kiwisolver==1.4.9
lightgbm>=4.5.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib>=3.8.2
maxminddb==2.8.2
multidict==6.7.0
narwhals==2.8.0
numpy>=1.26.0
oauthlib==3.3.1
packaging==25.0
pandas>=2.1.3
passlib==1.7.4
pillow==11.3.0
plotly==6.3.1
postgrest==1.1.1
prometheus-fastapi-instrumentator==6.1.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic>=2.7.0,<3.0.0
pydantic-settings==2.10.1

PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
realtime==2.6.0
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn>=1.3.2
scipy>=1.11.0
seaborn==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.27.0
storage3==0.12.0
StrEnum==0.4.15
stripe==12.4.0
supabase==2.17.0
supafunc==0.10.1
threadpoolctl==3.6.0
torch==2.9.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
webencodings==0.5.1
websockets==15.0.1
xgboost==2.0.2
yarl==1.22.0
